% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{char_tokenizer}
\alias{char_tokenizer}
\title{Returns vector of tokens, tokenized by character}
\usage{
char_tokenizer(input)
}
\arguments{
\item{input}{String to tokenize}
}
\value{
vector of characters
}
\description{
Returns vector of tokens, tokenized by character
}
\examples{
char_tokenizer('hello')
# c("h", "e", "l", "l", "o")
}
