% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{cvc}
\alias{cvc}
\title{Function returns TRUE if and only if there is a sequence of cvc at the end of the word left after you remove suffix y1 from the word x1
except when the second consonant is w, x or y}
\usage{
cvc(x1, y1 = "")
}
\arguments{
\item{x1}{word}

\item{y1}{suffix}
}
\description{
Function returns TRUE if and only if there is a sequence of cvc at the end of the word left after you remove suffix y1 from the word x1
except when the second consonant is w, x or y
}
