% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{d}
\alias{d}
\title{Function returns TRUE if and only if there is a double consonant at the end of the word left after you remove suffix y1 from the word x1}
\usage{
d(x1, y1 = "")
}
\arguments{
\item{x1}{word}

\item{y1}{suffix}
}
\description{
Function returns TRUE if and only if there is a double consonant at the end of the word left after you remove suffix y1 from the word x1
}
