% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{filt}
\alias{filt}
\title{filt combinator which returns filtered vector by predicate}
\usage{
filt(x, p)
}
\arguments{
\item{x}{vector to filter}

\item{p}{predicate}
}
\value{
filtered vector
}
\description{
filt combinator which returns filtered vector by predicate
}
\examples{
filt(c(1, 2, 3, 4, 5, 6), (~ x > 3))
}
