% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{flatMap}
\alias{flatMap}
\title{flatMap combinator}
\usage{
flatMap(x, f)
}
\arguments{
\item{x}{list of vectors to which apply}

\item{f}{function (or lambda expression) for mapping}
}
\value{
vector with applyed function
}
\description{
flatMap combinator
}
\examples{
flatMap(list(c(1, 2, 3, 4), c(5, 6), c(7, 8, 9)), (~ x^2))
}
