% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{flatten}
\alias{flatten}
\title{flatten converts list of vectors to vector}
\usage{
flatten(x)
}
\arguments{
\item{x}{list of vectors}
}
\value{
vector with applyed function
}
\description{
The same as x %flatMap% identity
}
\examples{
flatten(list(c(1, 2, 3, 4), c(5, 6), c(7, 8, 9)))
}
