% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functor}
\alias{functor}
\title{Returns functor for 'map'-like applying}
\usage{
functor(init = c(0))
}
\arguments{
\item{init}{Initial vector}
}
\value{
Functor
}
\description{
Returns functor for 'map'-like applying
}
\examples{
functor(c(1, 2, 3, 4))()
# c(1, 2, 3, 4)
functor(c(1, 2, 3, 4))(function(x) {x * 2})()
# c(2, 4, 6, 8)
}
