% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{\%all\%}
\alias{\%all\%}
\title{all combinator which returns TRUE if all elements has TRUE predicate}
\usage{
x \%all\% p
}
\arguments{
\item{x}{vector to check}

\item{p}{predicate}
}
\value{
boolean
}
\description{
all combinator which returns TRUE if all elements has TRUE predicate
}
