% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{\%any\%}
\alias{\%any\%}
\title{any combinator which returns TRUE if any element has TRUE predicate}
\usage{
x \%any\% p
}
\arguments{
\item{x}{vector to check}

\item{p}{predicate}
}
\value{
boolean
}
\description{
any combinator which returns TRUE if any element has TRUE predicate
}
