% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{\%filt\%}
\alias{\%filt\%}
\title{filt combinator which returns filtered vector by predicate (infix version)}
\usage{
x \%filt\% p
}
\arguments{
\item{x}{vector to filter}

\item{p}{predicate}
}
\value{
filtered vector
}
\description{
filt combinator which returns filtered vector by predicate (infix version)
}
