% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{\%filt_not\%}
\alias{\%filt_not\%}
\title{filt_not combinator which returns filtered vector by reversed predicate (infix version)}
\usage{
x \%filt_not\% p
}
\arguments{
\item{x}{vector to filter}

\item{p}{predicate}
}
\value{
filtered vector
}
\description{
filt_not combinator which returns filtered vector by reversed predicate (infix version)
}
