% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{\%flatMap\%}
\alias{\%flatMap\%}
\title{flatMap combinator (infix version)}
\usage{
x \%flatMap\% f
}
\arguments{
\item{x}{list of vectors to which apply}

\item{f}{function (or lambda expression) for mapping}
}
\value{
vector with applyed function
}
\description{
flatMap combinator (infix version)
}
