% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{\%has\%}
\alias{\%has\%}
\title{has operator to check if vector contains element}
\usage{
x \%has\% e
}
\arguments{
\item{x}{vector to check}

\item{e}{element to check}
}
\value{
boolean
}
\description{
has operator to check if vector contains element
}
