% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{\%head_while\%}
\alias{\%head_while\%}
\title{head_while combinator which returns head elements while predicate is TRUE (infix version)}
\usage{
x \%head_while\% p
}
\arguments{
\item{x}{vector to filter}

\item{p}{predicate}
}
\value{
head filtered
}
\description{
head_while combinator which returns head elements while predicate is TRUE (infix version)
}
