% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{\%if\%}
\alias{\%if\%}
\title{if combinator which returns filtered vector by predicate as 'filter' (infix version)}
\usage{
x \%if\% p
}
\arguments{
\item{x}{vector to filter}

\item{p}{predicate}
}
\value{
filtered vector
}
\description{
if combinator which returns filtered vector by predicate as 'filter' (infix version)
}
