% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{\%if_not\%}
\alias{\%if_not\%}
\title{if_not combinator which returns filtered vector by reversed predicate as 'filt_not' (infix version)}
\usage{
x \%if_not\% p
}
\arguments{
\item{x}{vector to filter}

\item{p}{predicate}
}
\value{
filtered vector
}
\description{
if_not combinator which returns filtered vector by reversed predicate as 'filt_not' (infix version)
}
