% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{\%invoke_map\%}
\alias{\%invoke_map\%}
\title{invoke_map combinator which apply each function to each vector respectively (infix version)}
\usage{
x \%invoke_map\% f
}
\arguments{
\item{x}{list of vectos to which apply}

\item{f}{list of functions (or lambda expressions) for mapping}
}
\value{
list of vectors with applyed function
}
\description{
invoke_map combinator which apply each function to each vector respectively (infix version)
}
