% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{\%map\%}
\alias{\%map\%}
\title{map combinator (infix version)}
\usage{
x \%map\% f
}
\arguments{
\item{x}{vector to which apply}

\item{f}{function (or lambda expression) for mapping}
}
\value{
vector with applyed function
}
\description{
map combinator (infix version)
}
