% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{\%mapN\%}
\alias{\%mapN\%}
\title{mapN combinator (apply map for each subvector) (infix version)}
\usage{
x \%mapN\% f
}
\arguments{
\item{x}{list of vectos to which apply}

\item{f}{function (or lambda expression) for mapping}
}
\value{
list of vectors with applyed function
}
\description{
mapN combinator (apply map for each subvector) (infix version)
}
