% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{\%tail_while\%}
\alias{\%tail_while\%}
\title{tail_while combinator which returns tail elements while predicate is TRUE (infix version)}
\usage{
x \%tail_while\% p
}
\arguments{
\item{x}{vector to filter}

\item{p}{predicate}
}
\value{
tail filtered
}
\description{
tail_while combinator which returns tail elements while predicate is TRUE (infix version)
}
