% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{h1}
\alias{h1}
\title{Function that returns a logical vector of TRUE/FALSE depending on vowel/consonant
for instance, for the input 'cry', it returns 'FALSE' 'FALSE' 'TRUE'}
\usage{
h1(x)
}
\arguments{
\item{x}{vector of booleans}
}
\description{
Function that returns a logical vector of TRUE/FALSE depending on vowel/consonant
for instance, for the input 'cry', it returns 'FALSE' 'FALSE' 'TRUE'
}
