% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{head_while}
\alias{head_while}
\title{head_while combinator which returns head elements while predicate is TRUE}
\usage{
head_while(x, p)
}
\arguments{
\item{x}{vector to filter}

\item{p}{predicate}
}
\value{
head filtered
}
\description{
head_while combinator which returns head elements while predicate is TRUE
}
\examples{
head_while(c(1, 2, 3, 4, 1, 2), (~ x < 3))
}
