% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{identity}
\alias{identity}
\title{identity function which returns element itselves}
\usage{
identity(x)
}
\arguments{
\item{x}{any object}
}
\value{
x parameter
}
\description{
identity function which returns element itselves
}
