% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{invoke_map}
\alias{invoke_map}
\title{invoke_map combinator which apply each function to each vector respectively}
\usage{
invoke_map(x, f)
}
\arguments{
\item{x}{list of vectos to which apply}

\item{f}{list of functions (or lambda expressions) for mapping}
}
\value{
list of vectors with applyed function
}
\description{
invoke_map combinator which apply each function to each vector respectively
}
\examples{
invoke_map(list(c(1, 2, 3), c(4, 5, 6), c(7, 8, 9)), list((~ x^2), (~ x^3), (~ x^4)))
}
