% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{lambda_to_func}
\alias{lambda_to_func}
\title{Converts special defined lambda expression format to R function}
\usage{
lambda_to_func(expr)
}
\arguments{
\item{expr}{formula to convert}
}
\value{
Function for applying
}
\description{
Converts special defined lambda expression format to R function
}
\examples{
lambda_to_func(x ~ {
    a <- x + 1
    return (a * 2)
})(3)
# 8
}
