% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{map}
\alias{map}
\title{map combinator}
\usage{
map(x, f)
}
\arguments{
\item{x}{vector to which apply}

\item{f}{function (or lambda expression) for mapping}
}
\value{
vector with applyed function
}
\description{
map combinator
}
\examples{
map(c(1, 2, 3, 4), (~ x^2))
# c(1, 4, 9, 16)
}
