% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{mapN}
\alias{mapN}
\title{mapN combinator (apply map for each elements of subvectors respectively)}
\usage{
mapN(x, f)
}
\arguments{
\item{x}{list of vectos to which apply}

\item{f}{function (or lambda expression) for mapping}
}
\value{
vector with applyed function
}
\description{
mapN combinator (apply map for each elements of subvectors respectively)
}
\examples{
mapN(list(c(1, 2, 3), c(4, 5, 6), c(7, 8, 9)), (x ~ y ~ z ~ x + y + z))
# list(c(1, 4, 9), c(16, 25, 36), c(49, 64))
}
