% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{nGramm}
\alias{nGramm}
\title{Returns n-gramms for given vector of some elements}
\usage{
nGramm(tokens, n)
}
\arguments{
\item{tokens}{Vector of elements to be n-grammed}

\item{n}{Size of n-gramm}
}
\value{
List of vector of n-grammed elements
}
\description{
Returns n-gramms for given vector of some elements
}
\examples{
vec <- c(1, 2, 3, 4, 5)
nGramm(vec, 3)
}
