% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{PorterStemmer.en}
\alias{PorterStemmer.en}
\alias{porterStem.en}
\title{Returns stemmed word}
\usage{
porterStem.en(x)
}
\arguments{
\item{x}{Input word to be stemmed}
}
\value{
Stemmed word
}
\description{
Returns stemmed word
}
\examples{
porterStem.en('actually')
}
