% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{PorterStemmer.ru}
\alias{PorterStemmer.ru}
\alias{porterStem.ru}
\title{Returns stemmed word}
\usage{
porterStem.ru(word)
}
\arguments{
\item{word}{Input word to be stemmed}
}
\value{
Stemmed word
}
\description{
Returns stemmed word
}
\examples{
porterStem.ru('РїСЂРёР»РµР¶РЅРѕСЃС‚СЊ')
}
