% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{space_tokenizer}
\alias{space_tokenizer}
\title{Returns vector of tokens, tokenization by space}
\usage{
space_tokenizer(input, sep = " ")
}
\arguments{
\item{input}{String to tokenize}

\item{sep}{Separator to split (default - space)}
}
\value{
vector of words
}
\description{
Returns vector of tokens, tokenization by space
}
