% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{symbolNGramm}
\alias{symbolNGramm}
\title{Returns n-gramms of characters for given input string}
\usage{
symbolNGramm(input, n, strings = TRUE)
}
\arguments{
\item{input}{Input string}

\item{n}{Size of n-gramm}

\item{strings}{Boolean - should function return vector of string n-gramms. FALSE - return list of vectors of characters}
}
\value{
Vector of strings or List of vector of characters @seealso strings
}
\description{
Returns n-gramms of characters for given input string
}
\examples{
symbolNGramm('hello world', 4)
}
