% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{v}
\alias{v}
\title{Function returns TRUE if and only if there is a vowel in the word left after you remove suffix y1 from the word x1}
\usage{
v(x1, y1 = "")
}
\arguments{
\item{x1}{word}

\item{y1}{suffix}
}
\description{
Function returns TRUE if and only if there is a vowel in the word left after you remove suffix y1 from the word x1
}
