% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{word_tokenizer}
\alias{word_tokenizer}
\title{Returns vector of tokens, tokenized by word}
\usage{
word_tokenizer(input)
}
\arguments{
\item{input}{String to tokenize}
}
\value{
vector of words
}
\description{
Returns vector of tokens, tokenized by word
}
\examples{
word_tokenizer('hello world and etc.')
# c("hello", "world", "and", "etc")
}
