% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{wordNGramm}
\alias{wordNGramm}
\title{Returns n-gramms of words for given input string}
\usage{
wordNGramm(input, n, clearData = TRUE)
}
\arguments{
\item{input}{Input string}

\item{n}{Size of n-gramm}

\item{clearData}{Boolean - should function clear input data (remove all useless characters)}
}
\value{
List of vector of words
}
\description{
Returns n-gramms of words for given input string
}
\examples{
wordNGramm('Never gonna give you up Never gonna let you down', 3)
}
